requirements.txt
python-telegram-bot
aiogram==2.25.2
python-dotenv>=1.0
firebird-driver==1.9.0
transformers>=4.30
torch
transformers
apscheduler
openpyxl
matplotlib
beautifulsoup4
pyautogui 
keyboard 
pygetwindow
